# doc/api.yaml
openapi: 3.0.0
info:
  title: wasaText
  version: 1.0.0
  description: API per il progetto wasaText
servers:
  - url: http://127.0.0.1:3000
    description: Server di sviluppo in locale

  # This YAML file defines a set of API operations for a messaging application.
  # Below is a brief description of each operation:
  #
  # doLogin:
  # x - Simplified login operation for user authentication.
  #
  # setMyUserName:
  # x - Allows the user to set or update their username.
  #
  # getMyConversations:
  # x - Retrieves a list of all conversations associated with the user.
  #
  # getConversation:
  # x - Fetches the details of a specific conversation by its identifier.
  #
  # sendMessage:
  # x - Sends a message to a specific conversation or user.
  #
  # forwardMessage:
  # x - Forwards an existing message to another conversation or user.
  #
  # commentMessage:
  # x - Adds a comment to a specific message.
  #
  # uncommentMessage:
  # x - Removes a previously added comment from a message.
  #
  # deleteMessage:
  # x - Deletes a specific message from a conversation.
  #
  # addToGroup:
  # x - Adds a user to a group conversation.
  #
  # leaveGroup:
  # x - Allows the user to leave a group conversation.
  #
  # setGroupName:
  #   - Updates the name of a group conversation.
  #
  # setMyPhoto:
  # x - Sets or updates the user's profile photo.
  #
  # setGroupPhoto:
  #   - Sets or updates the photo of a group conversation.

paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Maria
                  pattern: "ˆ.*?$"
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        "201":
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: string
                    example: "abcdef012345"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid username"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"

  /user:
    get:
      tags:
        - utente
      operationId: getListOfUsers
      summary: Ottieni la lista di utenti
      description: >
        Ottiene la lista di utenti registrati.
        Restituisce un elenco di utenti con i loro ID e nomi.
      responses:
        "200":
          description: Elenco degli utenti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "500":
          description: Errore interno del server
  /me/username:
    put:
      tags:
        - utente
      operationId: setMyUserName
      summary: Imposta il nome utente
      description: >
        Imposta il nome utente per l'utente corrente.
        Se il nome utente è già in uso, restituisce un errore.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 16
              required:
                - username
      responses:
        "200":
          description: Nome utente impostato correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Nome utente non valido
        "409":
          description: Nome utente già in uso
        "500":
          description: Errore interno del server
    get:
      tags:
        - utente
      operationId: getMyUserName
      summary: Ottieni il nome utente
      security:
        - bearerAuth: []
      description: >
        Ottiene il nome utente per l'utente corrente.
        Restituisce l'ID e il nome dell'utente.
      responses:
        "200":
          description: Nome utente ottenuto correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: "marta"
        "404":
          description: Utente non trovato
        "500":
          description: Errore interno del server
  /me/photo:
    post:
      tags:
        - utente
      operationId: setMyPhoto
      summary: Imposta la foto del profilo
      description: >
        Imposta la foto del profilo per l'utente corrente.
        Restituisce l'ID della foto impostata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
              required:
                - photo
      responses:
        "200":
          description: Foto del profilo impostata correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  photoID:
                    type: integer
                    example: 1
        "400":
          description: Foto non valida
        "500":
          description: Errore interno del server
  /user/{userID}:
    parameters:
      - $ref: "#/components/parameters/userID"

    get:
      tags:
        - utente
      operationId: getUser
      summary: Ottieni informazioni sull'utente
      description: >
        Ottiene le informazioni sull'utente specificato dall'ID.
        Restituisce l'ID e il nome dell'utente.
      responses:
        "200":
          description: Informazioni sull'utente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Utente non trovato
        "500":
          description: Errore interno del server

  /conversation:
    get:
      tags:
        - utente
      operationId: getMyConversations
      summary: Ottieni le conversazioni dell'utente
      description: >
        Ottiene tutte le conversazioni associate all'utente specificato.
        Restituisce un elenco di conversazioni.

      responses:
        "200":
          description: Elenco delle conversazioni
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        "404":
          description: Utente non trovato
        "500":
          description: Errore interno del server

  /conversation/{conversationID}:
    parameters:
      - $ref: "#/components/parameters/conversationID"
    get:
      tags:
        - conversazione
      operationId: getConversations
      summary: Ottieni una conversazione specifica
      description: >
        Ottiene i dettagli di una conversazione specifica.
        Restituisce l'ID della conversazione e i messaggi associati.
      responses:
        "200":
          description: Dettagli della conversazione
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Conversazione non trovata
        "500":
          description: Errore interno del server

  /conversation/{conversationID}/message:
    parameters:
      - $ref: "#/components/parameters/conversationID"

    get:
      tags:
        - conversazione
      operationId: getMessages
      summary: Ottieni i messaggi di una conversazione
      description: >
        Ottiene tutti i messaggi di una conversazione specifica.
        Restituisce un elenco di messaggi.
      responses:
        "200":
          description: Elenco dei messaggi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "404":
          description: Conversazione non trovata
        "500":
          description: Errore interno del server

    post:
      tags:
        - conversazione
      operationId: sendMessage
      summary: Invia un messaggio
      description: >
        Invia un messaggio a una conversazione specifica.
        Restituisce l'ID del messaggio inviato.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
      responses:
        "200":
          description: Messaggio inviato correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  messageID:
                    type: integer
                    example: 1
        "400":
          description: Messaggio non valido
        "404":
          description: Conversazione non trovata
        "500":
          description: Errore interno del server

  /conversation/{conversationID}/message/{messageID}:
    parameters:
      - $ref: "#/components/parameters/conversationID"
      - $ref: "#/components/parameters/messageID"

    post:
      tags:
        - conversazione
      operationId: forwardMessage
      summary: Inoltra un messaggio
      description: >
        Inoltra un messaggio a una conversazione specifica.
        Restituisce l'ID del messaggio inoltrato.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationID:
                  type: integer
                  example: 1
              required:
                - conversationID
      responses:
        "200":
          description: Messaggio inoltrato correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  messageID:
                    type: integer
                    example: 1
        "400":
          description: Messaggio non valido
        "404":
          description: Conversazione non trovata
        "500":
          description: Errore interno del server

  /messages/{messageID}:
    parameters:
      - $ref: "#/components/parameters/messageID"

    get:
      tags:
        - messaggio
      operationId: getMessage
      summary: Ottieni un messaggio specifico
      description: >
        Ottiene i dettagli di un messaggio specifico.
        Restituisce l'ID del messaggio e il contenuto.
      responses:
        "200":
          description: Dettagli del messaggio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Messaggio non trovato
        "500":
          description: Errore interno del server
    delete:
      tags:
        - messaggio
      operationId: deleteMessage
      summary: Elimina un messaggio
      description: >
        Elimina un messaggio specifico.
        Restituisce l'ID del messaggio eliminato.
      responses:
        "200":
          description: Messaggio eliminato correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  messageID:
                    type: integer
                    example: 1
        "404":
          description: Messaggio non trovato
        "500":
          description: Errore interno del server

  /messages/{messageID}/comment:
    parameters:
      - $ref: "#/components/parameters/messageID"

    post:
      tags:
        - messaggio
      operationId: commentMessage
      summary: Aggiungi un commento a un messaggio
      description: >
        Aggiunge un commento a un messaggio specifico.
        Restituisce l'ID del commento aggiunto.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 500
              required:
                - content
      responses:
        "200":
          description: Commento aggiunto correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  commentID:
                    type: integer
                    example: 1
        "400":
          description: Commento non valido
        "404":
          description: Messaggio non trovato
        "500":
          description: Errore interno del server

  /messages/{messageID}/comment/{commentID}:
    parameters:
      - $ref: "#/components/parameters/messageID"
      - $ref: "#/components/parameters/commentID"
    delete:
      tags:
        - messaggio
      operationId: uncommentMessage
      summary: Rimuovi un commento da un messaggio
      description: >
        Rimuove un commento da un messaggio specifico.
        Restituisce l'ID del commento rimosso.
      responses:
        "200":
          description: Commento rimosso correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  commentID:
                    type: integer
                    example: 1
        "404":
          description: Messaggio o commento non trovato
        "500":
          description: Errore interno del server
  /group:
    post:
      tags:
        - gruppo
      operationId: createGroup
      summary: Crea un nuovo gruppo
      description: >
        Crea un nuovo gruppo di lavoro.
        Restituisce l'ID del gruppo creato.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
              required:
                - name
      responses:
        "200":
          description: Gruppo creato correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  groupID:
                    type: integer
                    example: 1
        "400":
          description: Nome del gruppo non valido
        "500":
          description: Errore interno del server

  /group/{groupID}:
    parameters:
      - $ref: "#/components/parameters/groupID"

    get:
      tags:
        - gruppo
      operationId: getGroup
      summary: Ottieni informazioni su un gruppo
      description: >
        Ottiene le informazioni su un gruppo specifico.
        Restituisce l'ID e il nome del gruppo.
      responses:
        "200":
          description: Informazioni sul gruppo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Gruppo non trovato
        "500":
          description: Errore interno del server

  /group/{groupID}/photo:
    parameters:
      - $ref: "#/components/parameters/groupID"

    post:
      tags:
        - gruppo
      operationId: setGroupPhoto
      summary: Imposta la foto del gruppo
      description: >
        Imposta la foto del gruppo specificato.
        Restituisce l'ID della foto impostata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
              required:
                - photo
      responses:
        "200":
          description: Foto del gruppo impostata correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  photoID:
                    type: integer
                    example: 1
        "400":
          description: Foto non valida
        "404":
          description: Gruppo non trovato
        "500":
          description: Errore interno del server
  /group/{groupID}/name:
    parameters:
      - $ref: "#/components/parameters/groupID"

    get:
      tags:
        - gruppo
      operationId: getGroupName
      summary: Ottieni il nome di un gruppo
      description: >
        Ottiene il nome di un gruppo specifico.
        Restituisce l'ID e il nome del gruppo.
      responses:
        "200":
          description: Nome del gruppo
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: "gruppo di lavoro"
        "404":
          description: Gruppo non trovato
        "500":
          description: Errore interno del server
    post:
      tags:
        - gruppo
      operationId: setGroupName
      summary: Imposta il nome del gruppo
      description: >
        Imposta il nome del gruppo specificato.
        Restituisce l'ID del gruppo aggiornato.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
              required:
                - name
      responses:
        "200":
          description: Nome del gruppo impostato correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  groupID:
                    type: integer
                    example: 1
        "400":
          description: Nome del gruppo non valido
        "404":
          description: Gruppo non trovato
        "500":
          description: Errore interno del server
  /group/{groupID}/members:
    parameters:
      - $ref: "#/components/parameters/groupID"

    get:
      tags:
        - gruppo
      operationId: getGroupMembers
      summary: Ottieni i membri di un gruppo
      description: >
        Ottiene tutti i membri di un gruppo specifico.
        Restituisce un elenco di membri del gruppo.
      responses:
        "200":
          description: Elenco dei membri del gruppo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "404":
          description: Gruppo non trovato
        "500":
          description: Errore interno del server
    post:
      tags:
        - gruppo
      operationId: addToGroup
      summary: Aggiungi un membro a un gruppo
      description: >
        Aggiunge un membro a un gruppo specifico.
        Restituisce l'ID del membro aggiunto.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userID:
                  type: integer
                  example: 1
              required:
                - userID
      responses:
        "200":
          description: Membro aggiunto correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  memberID:
                    type: integer
                    example: 1
        "400":
          description: ID utente non valido
        "404":
          description: Gruppo non trovato
        "500":
          description: Errore interno del server

  /group/{groupID}/members/{userID}:
    parameters:
      - $ref: "#/components/parameters/groupID"
      - $ref: "#/components/parameters/userID"

    delete:
      tags:
        - gruppo
      operationId: leaveGroup
      summary: Lascia un gruppo
      description: >
        Permette a un membro di lasciare un gruppo specifico.
        Restituisce l'ID del membro rimosso.
      responses:
        "200":
          description: Membro rimosso correttamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  memberID:
                    type: integer
                    example: 1
        "404":
          description: Gruppo o utente non trovato
        "500":
          description: Errore interno del server

components:
  #schemi per definire oggetti
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "marta"
    Group:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "gruppo di lavoro"
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
    Conversation:
      type: object
      properties:
        id:
          type: integer
          example: 1
        Group:
          $ref: "#/components/schemas/Group"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    Message:
      type: object
      properties:
        sender:
          type: string
          example: "marta"
        content:
          type: string
          example: "Ciao a tutti!"
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        conversationId:
          type: integer
          example: 1

  #parametri utilizzati come richieste dell'utente
  parameters:
    username:
      name: username
      in: path
      required: true
      schema:
        type: string
        minLength: 3
        maxLength: 16
      description: il nome dell'utente

    userID:
      name: userID
      in: path
      required: true
      schema:
        type: string
      description: l'identificazione univoca dell'utente

    groupID:
      name: groupID
      in: path
      required: true
      schema:
        type: string
      description: l'identificazione univoca del gruppo

    conversationID:
      name: conversationID
      in: path
      required: true
      schema:
        type: string
      description: l'identificazione univoca della conversazione

    messageID:
      name: messageID
      in: path
      required: true
      schema:
        type: integer
      description: l'identificazione univoca del messaggio

    commentID:
      name: commentID
      in: path
      required: true
      schema:
        type: integer
      description: l'identificazione univoca del commento

  #risposte di errore predefinite
  responses:
    ErrorResponse:
      description: risposta di errore
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 404
              message:
                type: string
                example: "element not found"

security:
  - bearerAuth: [] # use the same name as above